cart item update fixed

+++ apps/orders/api_endpoints/Order/Order_Create/views.py
@@ -1,19 +1,37 @@
 from django.db.models.aggregates import Sum
+from drf_yasg import openapi
+from drf_yasg.utils import swagger_auto_schema
 from rest_framework.generics import CreateAPIView
-from rest_framework.permissions import IsAuthenticated
 
 from apps.orders.models import OrderItem
+from apps.cart.models import CartItem
 from .serializers import OrderCreateSerializer
 
 
 class OrderCreateAPIView(CreateAPIView):
     serializer_class = OrderCreateSerializer
 
+    @swagger_auto_schema(
+        manual_parameters=[
+            openapi.Parameter('device_id', openapi.IN_QUERY, description='Device id', type=openapi.TYPE_STRING),
+        ]
+    )
+    def post(self, request, *args, **kwargs):
+        return self.create(request, *args, **kwargs)
+
     def perform_create(self, serializer):
-        user = self.request.user
-        final_price = user.cart.items.aggregate(final_price=Sum('cost'))['final_price']
+        user = self.request.user if self.request.user.is_authenticated else None
+        device_id = self.request.query_params.get('device_id', None)
+        if user and device_id:
+            raise ValueError('device_id is needed only for guest users.')
+        if user:
+            items = CartItem.objects.filter(cart__user=user)
+        else:
+            items = CartItem.objects.filter(device_id=device_id)
+
+        final_price = items.aggregate(final_price=Sum('cost'))['final_price']
         order = serializer.save(user=user, final_price=final_price)
-        for item in user.cart.items.all():
+        for item in items:
             OrderItem.objects.create(order=order, product=item.product, quantity=item.quantity, cost=item.cost)
             item.delete()
